+++ packages/apptive_grid_core/lib/network/apptive_grid_client.dart
@@ -30,6 +30,7 @@ class ApptiveGridClient {
   /// Close the connection on the httpClient
   void dispose() {
     _client.close();
+    _authenticator.dispose();
   }
 
   /// Headers that are used for multiple Calls