+++ packages/apptive_grid_core/test/mocks.dart
@@ -5,9 +5,10 @@ import 'package:apptive_grid_core/network/io_authenticator.dart'
 
 import 'package:http/http.dart';
 import 'package:mocktail/mocktail.dart';
-import 'package:openid_client/openid_client.dart' as openid;
+import 'package:openid_client_fork/openid_client.dart' as openid;
 import 'package:url_launcher_platform_interface/url_launcher_platform_interface.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
+import 'package:uni_links_platform_interface/uni_links_platform_interface.dart';
 
 class MockHttpClient extends Mock implements Client {}
 
@@ -23,6 +24,10 @@ class MockToken extends Mock implements openid.TokenResponse {}
 class MockAuthClient extends Mock implements openid.Client {}
 
 class MockAuthenticator extends Mock implements Authenticator {}
+
+class MockUniLinks extends Mock
+    with MockPlatformInterfaceMixin
+    implements UniLinksPlatform {}
 
 class MockUrlLauncher extends Mock
     with MockPlatformInterfaceMixin